choicer/requirements.txt
django==2.2
Pillow==7.0
factoryboy==2.12.0
coverage==5.0.3